+++ CHANGELOG.md
@@ -320,3 +320,4 @@ Length two tests rewritten for 1.6 and TestGraph.
 MakeRdfTermFromSourceTestCase added with five cases.  prepare_columns refactored to return rdflib terms.  
 __do_the_update refactored to used rdflib terms.  Simplifies code, handles all rdflib term types and compares.  All
 test cases pass.  Removing VIVO date dependencies from PumpUpdateDataTestCases.  Two cases to improve.
+* **2015-09-20** Begin examples/grants